Uploading Makefile
A Makefile in C++ is a script that defines a set of rules and instructions for compiling and building a C++ program. It specifies how the source code files should be compiled and linked to create an executable. Makefiles are used to automate the build process and manage dependencies among source files.

This is generally useful for Linux users.