Uploading Makefile
A Makefile in C++ is a script that defines a set of rules and instructions for compiling and building a C++ program. It specifies how the source code files should be compiled and linked to create an executable. Makefiles are used to automate the build process and manage dependencies among source files.

This is generally useful for Linux users.

Updated as per the requirement of the question.

-> -Wall: Used for debugging as it prints all the warnings for the file

-> -O2: Used for optimized execution of the code

Ref.
1) https://stackoverflow.com/a/655390/10867044
2) https://youtu.be/6Gw1rNyTJWA?si=OlMx4ckR-yveVfhM
3) https://www.geeksforgeeks.org/compiling-with-g-plus-plus/